created new component folders (with presentational and container types) and moved 'App' to the container component
