+++ src/components/container/App/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './App.css';
+
+function App({children}) {
+  return (
+    <div className="App">
+      {children}
+    </div>
+  );
+}
+
+export default App;
